+++ documentation/changelog.md
@@ -1,2 +1,3 @@
 # Parte 3
 - Planeta agora tem massa
+- Atuacao tem um ID sintetico